backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
firebase-admin==6.6.0
google-generativeai==0.8.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
ffmpeg-python==0.2.0
python-dotenv==1.0.1
httpx==0.28.1
openai==1.58.1
